requirements.txt
flask
gunicorn
twilio
langchain
langchain-community
langchain-huggingface
langchain-groq
langchain-core
langchain-chroma
chromadb
sentence-transformers
python-dotenv
streamlit
